modify models and rename DbModels folder to Models
