+++ ValantDemoApi/ValantDemoApi/Models/Game.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ValantDemoApi.DbModels
+namespace ValantDemoApi.Models
 {
   public class Game
   {
     public int Id { get; set; }
     public int MazeId { get; set; }
-    public Tuple<int, int> CurrentPosition {get; set;}
+    //public Cell CurrentPosition {get; set;}
   }
 }
